int8_t for error returning and status

+++ include/network/network_p2p.h
@@ -10,19 +10,19 @@ struct network_p2p {
 };
 
 void network_p2p_data_init(struct network_p2p *, struct network_conf,
-        int (*)(const struct string_st *, struct string_st *),
-        int (*)(const struct string_st *));
+                           int8_t (*)(const struct string_st *, struct string_st *),
+                           int8_t (*)(const struct string_st *));
 void network_p2p_data_free(struct network_p2p *);
 
 void network_p2p_start(struct network_p2p *);
 void network_p2p_close(struct network_p2p *);
 
-int network_p2p_set_hosts(struct network_p2p *, const struct string_st *);
+int8_t network_p2p_set_hosts(struct network_p2p *, const struct string_st *);
 void network_p2p_get_hosts(const struct network_p2p *, struct string_st *);
-int network_p2p_connected(struct network_p2p *);
+int8_t network_p2p_connected(struct network_p2p *);
 
-int network_p2p_get(struct network_p2p *, const struct string_st *, struct string_st *);
-int network_p2p_send(struct network_p2p *, const struct string_st *);
+int8_t network_p2p_get(struct network_p2p *, const struct string_st *, struct string_st *);
+int8_t network_p2p_send(struct network_p2p *, const struct string_st *);
 
 
 #endif //NETWORK_P2P_H
